# api/swagger.yaml
---
swagger: "2.0"
info:
  description: "A simple API to access the data and services of the project GeohistoricalData.\
    \ Find out more at [http://geohistoricaldata.org/api/docs](http://geohistoricaldata.org/api/docs)"
  version: "1.0.0"
  title: "GeohistoricalData API"
  termsOfService: ""
  contact:
    email: "contact@geohistoricaldata.org"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:8080"
basePath: "/"
tags:
- name: "geocoding"
  description: "Geocode a set of pairs <Address, Date>."
  externalDocs:
    description: "Find out more"
    url: "http://api.geohistoricaldata.org/docs"
schemes:
- "http"
paths:
  /geocoding:
    get:
      tags:
      - "Geocoding"
      summary: "Geocode an adress at a given date."
      description: "Some description."
      operationId: "geocodeGET"
      parameters:
        - in: "query"
          name: "address"
          type: "string"
          default: "12 rue du temple"
        - in: "query"
          name: "date"
          type: "integer"
          default: 1850
        - in: "query"          
          name: precision
          type: "boolean"
          default: true
        - in: query
          name: "maxresults"
          type: "integer"
          default: 1
      responses:
        200:
          description: "Success"
        400:
          description: "Bad request"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "Geocoding"
    post:
      tags:
      - "Geocoding"
      summary: "Geocode a set of addresses at a given date."
      description: "Some description."
      operationId: "geocodePOST"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        required: false
        schema:
          type: "array"
          items:
            $ref: "#/definitions/GeocodingQuery"
      responses:
        200:
          description: "Success"
        400:
          description: "Bad request"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "Geocoding"
definitions:
  Error:
    required:
    - "code"
    - "message"
    properties:
      code:
        type: "integer"
        format: "int32"
      message:
        type: "string"

  GeocodingQuery:
    type: "object"
    required:
    - "address"
    - "date"
    properties:
      address:
        type: "string"
        example: "1 rue du Temple, Paris"
      date:
        type: "integer"
        example: 1850
      precision:
        type: "boolean"
        example: true
        xml:
          name: "precise-geocoding"
        default: true
      maxresults:
        type: "integer"
        example: 1
        xml:
          name: "max-results"
    xml:
      name: "geohistorical-object"